Provide a way to serve packages from several directories
Using the packages_location of the local_apt_repository provider or the node['local_apt']['packages_locations'] and the default location.
This will serve all the packages contained in the directories and sub directories provided.

Remove the update action in favor of regenerating the repository on each run.
This will require apt-updating each time.